# apis/types/search_result_item.go
// Code generated by go-swagger; DO NOT EDIT.

package types

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/swag"
)

// SearchResultItem search result item in search results.
// swagger:model SearchResultItem

type SearchResultItem struct {

	// description just shows the description of this image
	Description string `json:"description,omitempty"`

	// is_automated means whether this image is automated.
	IsAutomated bool `json:"is_automated,omitempty"`

	// is_official shows if this image is marked official.
	IsOfficial bool `json:"is_official,omitempty"`

	// name represents the name of this image
	Name string `json:"name,omitempty"`

	// star_count refers to the star count of this image.
	StarCount int64 `json:"star_count,omitempty"`
}

/* polymorph SearchResultItem description false */

/* polymorph SearchResultItem is_automated false */

/* polymorph SearchResultItem is_official false */

/* polymorph SearchResultItem name false */

/* polymorph SearchResultItem star_count false */

// Validate validates this search result item
func (m *SearchResultItem) Validate(formats strfmt.Registry) error {
	var res []error

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

// MarshalBinary interface implementation
func (m *SearchResultItem) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *SearchResultItem) UnmarshalBinary(b []byte) error {
	var res SearchResultItem
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}